🚨 Errors: restructured errors.
- errors are now structured based on functional segments (repositories, etc).
- defined new errors for the collection respository layer.
