+++ internal/errors/repositories.go
@@ -1,0 +1,8 @@
+package errors
+
+// ------------- Collection Repository
+var (
+	ErrCollectionsRetrieval      = &TapaError{Code: 3000, Message: "Failed fetching all collections \n"}
+	ErrFoldersRetrieval          = &TapaError{Code: 3000, Message: "Failed fetching all folders \n"}
+	ErrRequestSummariesRetrieval = &TapaError{Code: 3000, Message: "Failed fetching all request summaries \n"}
+)
